backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
openai==2.21.0
Pillow>=10.0.0
geopy==2.4.1
pandas==2.2.3
python-dotenv==1.0.1
pydantic-settings==2.6.1
httpx==0.28.0
alembic==1.14.0
